logging: refine errors reporting
Define macro for exit code saying that user canceled the operation. We
want to be able to distinguish between fatal errors and user requests
because we want to react in different ways in frontends.

The code is 69 and there is no special reason for it. (except it must be
in range 0-255)

The current implementation allows programmers to set a global variable
xfunc_error_retval to any value nevertheless the new function hiding the
global variable was added in order to support bindings to other
languages.

Related to #146

Signed-off-by: Jakub Filak <jfilak@redhat.com>