wizard: eliminate evd->error_msg member
There is only one small user-visible change in this patch:
"...You can try another operation if available" part of
"Processing failed" message is gone.
This addition was shown only in expert mode.
I feel it is not needed since GUI workflow
makes it obvious anyway that another event can be selected and run.

Signed-off-by: Denys Vlasenko <dvlasenk@redhat.com>
Signed-off-by: Jiri Moskovcak <jmoskovc@redhat.com>