rhbz: don't pass NULL in platform argument
The platform field is not required. Server uses the default value if
the platform is empty.

From my point of view, it is much better to let user report a bug
instead of dieing due to a missing file (or worse crashing). Ordinary
users cannot fix it and will submit the dump dir for emergency analysis.

Closes rhbz#963020

Signed-off-by: Jakub Filak <jfilak@redhat.com>
Signed-off-by: Martin Milata <mmilata@redhat.com>