ureport: always add BTHASH link to a report
Add an url pointing to the report via BTHASH. Use prefix "ABRT Server"
because FAF uses the same prefix in reported_to result. The views should
display only the last entry of a single prefix. Once user report the
crash to Bugzilla, the consecutive run of reporter-ureport will add a
new "ABRT Server" entry to reported_to element but now the url points to
the report via report's ID.

The new entry is added in order to allow users to view the report on
server even if the crash is unknown.

Related to rhbz#977029

Signed-off-by: Jakub Filak <jfilak@redhat.com>