XML parser: keep region specific translation texts
Our xml parser tries to always find a translated value. If no exact
region locale is found, the last found language will be used.

There is a bug in the algorithm which causing that the parser always
uses the last found language.

This patch adds a condition which prevents the parser from replacing of
the exact region locale by the last found language.

Closes rhbz#909536

Signed-off-by: Jakub Filak <jfilak@redhat.com>