Bugzilla: pass Bugzilla_token in all relevat XMP RPC calls
Affected functions:
    Bug.update
    Bug.add_attachement
    Bug.add_comment
    Bug.get
    Bug.comments
    Bug.create
    Bug.search
    User.logout

Not affected functions:
    Bugzilla.version

Instead of returning a cookie, the User.login call now returns a token
that clients must pass in the Bugzilla_token parameter to subsequent
RPC calls. If the token is not passed, Bugzilla will treat the RPC
call as unauthenticated and will not allow access to non-public data.

See
https://partner-bugzilla.redhat.com/docs/en/html/api/Bugzilla/WebService.html#LOGGING_IN
for more details.

Client scripts that access Red Hat Bugzilla via XML-RPC or JSON-RPC
and use login cookies for authentication must be updated to instead
remember the token received when logging in and pass that token back
to Bugzilla in subsequent RPC calls.

[http://post-office.corp.redhat.com/archives/bugzilla-list/2014-April/msg00005.html]

Signed-off-by: Jakub Filak <jfilak@redhat.com>