mailx: improve notification e-mail format
Make it more obvious who is the sender. Ensure that the notify*
configuration won't be overwritten by the configuration changes of the
mailx plugin. Distinguish between e-mails for a new crash and a new
occurrence.

The notify* configuration could be stored in a new configuration file
but such an approach seems a bit overengineered.

Resolves: rhbz#1093375

Signed-off-by: Jakub Filak <jfilak@redhat.com>