ureport: provide default URLs
We have two general use cases:
1. Completely anonymous uReporting
2. Authenticated uReporting

For good security reasons, we have to use a standalone server for each of
these use cases and this requirement complicates the configuration.
Users would have to change two options to switch between these two
configurations.

The configuration can be simplified by providing default URL for both of
them and single out the URL according to state of the "SSLClientAuth"
configuration option.

1. Completely anonymous uReporting
  - empty "SSLClientAuth"

2. Authenticated uReporting
  - "rhsm" or "rhsm-entitlement" in "SSLClientAuth"

Related to rhbz#1140224

Signed-off-by: Jakub Filak <jfilak@redhat.com>