ureport: include AuthDataItems if SSLClientAuth is configured
Forcing users to configure more than one configuration option to enable
Authenticated uReporting isn't good idea. So for sake of simplicity
reporter-ureport will add the AuthDataItems in to an uReport by default
if the SSLClientAuth is set to some value. Putting 'no' to the
IncludeAuthData configuration option, stops reporter-ureport from
including the AuthDataItems in uReports.

This commit also comments out AuthDataItems in ureport.conf in order to
prevent leaking of security sensitive data in case of a bug in the code
assigning a default value to IncludeAuthData variable (it would be a
real disaster to accidentally send the auth data items from Fedora
boxes)

Related to rhbz#1139557

Signed-off-by: Jakub Filak <jfilak@redhat.com>