upload: read credentials from environment variables
We use environment variables for the per-user-configuration.

Adding new configuration options to the configuration file for
credentials doesn't make sense because the credentials can be part of
the URL (everything is plain text in the configuration file). But it
does make sense to add a support for environment variables which are
used to propagate the configuration from the GUI where the URL field is
plain text but the password field cannot be read.

Related to rhbz#1066486

Signed-off-by: Jakub Filak <jfilak@redhat.com>