dump_dir: allow (semi)recursive locking
This patch only tries to mitigate the consequences of a bug in code
where someone tries to lock a dump directory while it is already locked
by the same process. This usually happens when a callee accepts a path
to directory and opens it on its own or when someone forgets to call
dd_unlock() or in all the unpredictable circumstance we usually have to
face in ABRT.

It is not possible to implement the lock counter using only a symbolic
link and file system functions, thus I've decided to put the
responsibility of unlocking to the first dd_lock() caller and disallow
the consecutive callers to unlock the dump directory.

Related to abrt/abrt#898

Signed-off-by: Jakub Filak <jfilak@redhat.com>