lib: allow creating root owned problem directories from problem data
Without this patch libreport sets the owner of new problem directory
according to FILENAME_UID. This approach is not sufficient because ABRT
has introduced PrivateReports that should ensure that all problem
directories are owned by root. So ABRT needs a way to tell libreport to
create the new problem directory with uid=0.

Signed-off-by: Jakub Filak <jfilak@redhat.com>