dd: don't try to close not opened dir fd
dd_do_open() checks if dd_fd is not negative but if it is negative the
function calls dd_close() which passes dd_fd to close() in any case.

This commit adds a check to dd_close() which ensures that the function
does not try to close negative dd_fd.

Uncovered by coverity.

Signed-off-by: Jakub Filak <jfilak@redhat.com>