dd: set owner to UID when creating dd from problem data
That is how the creating of a new dump directory from a problem data was
working before the patches adding the meta-data owner and the patches
dividing the dump directory owner to the FS owner of the dump directory
and the owner of the problem.

The created dump directory must be accessible to the uid from
FILENAME_UID but must have the requested FS owner.

Signed-off-by: Jakub Filak <jfilak@redhat.com>