dd: extend the scope of DD_DONT_WAIT_FOR_LOCK
The current implementation uses the flag only to ignore *unlocked or
broken* problems that misses some of the required files. If the dump
directory is locked by an existing process, dd_lock() waits until the
process unlocks it and that causes problems in UI.

Example:
1. abrt-hook-ccpp creates a new dump directory and goes to generate
core_backtrace, which blocks it for several seconds.
2. An user wants to get list of all problems from abrt-dbus.
3. abrt-dbus got stucked on the new problem locked by abrt-hook-ccpp.
4. D-Bus connection time-outs, abrt-dbus becomes unreachable.

This patch adds a new condition which breaks the loop wait for lock if
the problem directory is locked by existing process. All other cases
are already handled and dd_lock() either returns an error or steals the
lock file if the locking process seems to not exist. Transitional states
like when the locking process just unlocked the directory are handled too.

errno is set to EAGAIN that is used by standard functions when an
operation failed but it is possible that a next call can be successful.
It should inform the calling function that the failure is not fatal and
the function can silently continue.

Related: rhbz#1243791

Signed-off-by: Jakub Filak <jfilak@redhat.com>