ureport: improve curl's error messages
There were cases the 'post_state->errmsg' was empty and error message looks
like "Failed to upload uReport to the server 'https://localhost:12345/faf' with
curl:". This is kind of confusing because there is no information related to
curl error part.  In the case the 'post_state->errmsg' is empty, error message
without "with curl:" part will be printed. Also print information stored in
'post_state->curl_error_msg', if is not empty.

Now, in the case the 'post_state->errmsg' is empty, the error message may look
like this:
Failed to upload uReport to the server 'https://localhost:12345/faf'
Error: curl_easy_perform: Problem with the SSL CA cert (path? access rights?)

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>