uploader: allow empty username and password
Due to added the possibility to configure SSH keys the username and the
password may not be set. Also there is an another reason why we decided to
allow those options empty. The username and the password can be defined within
URL and if they are not provided either by conf file or within URL,
reporter-uploader will ask for that.

Related to rhbz#1261120

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>