dd: add function getting stat of item
ABRT Problems2 DBus service must not die on invalid data and must
provide users with accurate error messages.

I need to get the file system size of a dump directory element and
verify that the passed name is valid item name in a single call. I need
to be able to distinguish between non-existing items and empty items.

dd_get_item_size() returns 0 for non-existing items and dies if the
passed name is invalid.

dd_exist() dies if the passed name is invalid.

This patch adds a new function called dd_item_stat() which validates the
passed name and reports errors through return value.

Implementation of dd_get_item_size() is replaced with usage of the new
function.

dd_get_item_size() stops killing the process if the passed name is not
valid because this really useless approach. It can be done because the
callers already have to check the return value for errors.

Signed-off-by: Jakub Filak <jfilak@redhat.com>