bugzilla: don't report private problem as comment
Before this patch reporter-bugzilla ignored the Private report request
and added a public comment to a duplicate bug because it was assumed
that the duplicate comment cannot contain anything security sensitive.

There are two problems with it. The assumption is invalid because the
comment contains all one-line files including 'cmdline' and the reporter
might added something private to the bug description.

Bugzilla comments can be made private but not all users have rights to
do so. On the contrary, all users can set a group to a bug report.
Hence, this commit teaches reporter-bugzilla to ask the user if he/she
wants to open a new, private bug report and immediately close it as a
duplicate of the original or terminate the reporting. The tool will ask
the question only if the users wants to open a private report and a
duplicate bug report is found.

Resolves: #1279453

Signed-off-by: Jakub Filak <jfilak@redhat.com>
Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>