mailx: use problem report api to define an emais' content
If formatting file is defined, the reporter-mailx
uses the given file to create content of emails. Section summary is used
for email's subject. If no formatting file is defined, the default formatting
is used and subject can be redefined either by mailx.conf file or by env
variable.

Related to rhbz#1281312

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>