dd: add functions for opening dd item
In cases where libreport users don't want to build contents of a dump
dir element in memory and save to disk using dd_save_* functions, they
had to guess file name and take care of file attributes. Forcing users
to take of this is a security risk.

This commit introduces new functions that will create a file inside of
a dump directory with correct name and file attributes.

For simplicity, only read only mode and read-write mode are allowed.

The read-write mode cause removal of the original item element as we
must never use truncate mode because of hard link threat (libreport
code runs under privileged user, so libreport must avoid rewriting
files - the correct approach is to remove the old one and create the new
one).

Sometimes we need to be able write some data and immediately read them.
This can be done by opening the file, writing the contents, closing the
file and re-opening it for reading. However, if we need to split the
work into chunks, than this approach becomes quite expensive.

Signed-off-by: Jakub Filak <jfilak@redhat.com>