workflows: add workflow for adding data to existing case
Change behavior of -t parameter in reporter-rhtsupport.

If option -t without support case ID is used and problem data in DIR was never
reported to RHTSupport, you will ask to enter case ID to which you want to
upload the FILEs. If the data in DIR was reported, the case ID is obtained from
'reported_to' element in problem dir. In this case, you will be asked if you
want to attach the data to the support case or you want to change the support
case ID.

Related to #1395285

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>

Conflicts:
	po/POTFILES.in
	src/plugins/rhtsupport_event.conf
	src/workflows/Makefile.am