ureport: By default ignore unpackaged executables
When problem occurs in executable that is not packages, reporting with
reporter-ureport crashes, since it cannot find any package data.
It also does not make sense to try to report such problems.

Therefore this commit introduces simple check, which tests if the
problem came from packages executable. By default, reporting unpackaged
problems is disabled in ureport.conf. It can be turned on for all
problems with changing the config file or only to report one unpackaged
problem with using report-ureport -p.

Note: Using -p or enabling it in config file make the reporter-ureport
fail. It is for discussion how we should treat such problems in FAF.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>