augeas: Use generic augeas modules
Most of the changes are just replacement of current regexps with in-build modules
from augeas. The changes do not make any major changes in the behaviour of the lenses.

The only thing that is different is that Util.empty [1] module ignores empty comments.
This means that empty commented lines in the configuration files do not show up in
augeas tree and because of that they can't be changed using the augeas.

The goal of this PR is to simplify the augeas lenses but it should also fix the linked
bugzilla bug.

[1] http://augeas.net/docs/references/1.4.0/lenses/files/util-aug.html#Util.empty

Related: rhbz#1649742

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>