reportclient: Find and download required debuginfo packages
The current solution finds packages for given build-ids and then
downloads them. The problem is that some debuginfo packages require
other packages and if they are not available the generated backtrace
becomes unusable.

This commit adds a query for required packages and downloads them together
with the main packages.

Related: rhbz#1515265

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>