wizard: Do not always exclude from report
Before this commit excluding happened every time a new page was
displayed. Excluding should happen only after user had change to pick
what is going to be excluded.

Therefore we catch signal `map` which means that the table with details
for selecting what should be attached was displayed to user.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>