cli: run-command: Replace use of vfork() with fork()
vfork() replacing fork() was most likely an optimization, but, over
time, changes were added that violate the contract, i.e. the code
started going things other than exec and exit:
https://pubs.opengroup.org/onlinepubs/009695399/functions/vfork.html.

posix_spawn(), as recommended by Clang, would not be a suitable
replacement, as we need to call tcsetpgrp() on the child, doing which
after exec is too late if the goal is to avoid races.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>