cli: Continue running even if some events have no commands
Previously, if any event in the event chain (be it custom using '-e' or from a
workflow) had no commands specified, the processing of the chain would be
terminated with an error.

Following this change, the behaviour changes to only reporting that no
processing was done but continuing with the run.

This is reasonable since a workflow might specify a set of events to be run
using a wildcard, yet some of the matching events might not be usable for all
problem types.