gtk: Better handling of workflows with wildcarded events
1. Remove ad-hoc wildcard expansion code. Wildcards are already expanded
in the workflow parser.

2. Don't interrupt processing if an event has no actions specified that
match the current problem. This allows workflows with event wildcards to
work properly.

For instance, FedoraPython3 specifies 'collect_*' as one of its events.
Previously, this would (usually) fail on event collect_GConf, since out
of the box, it has actions specified only for problems of type CCpp and
with libgconf-2 loaded dynamically.