gtk: Fix infinite loop crashing the reporting
The problem was caused by a recursive spiral of death after a page
change in certain situations.

When transitioning to the progress page from a different one (for
instance after filling out a Bugzilla comment), the switch-page signal
handler on_page_prepare() is invoked. There, the next event's action is
spawned automatically by calling start_event_run() synchronously.

The problem arises if there are no actions to run. In that case,
start_event_run() calls on_next_btn_cb() synchronously to skip to the
next event in sequence, which again calls the on_page_prepare() handler.
Since we're still in the call stack of the switch-page signal handler,
the page hasn't changed yet. This causes on_page_prepare() to presume
we're about to change the page, thus spinning into next_btn_cb() again,
and so on, and so on *sniff*

This commit fixes this by selecting the next event with actions directly
in the signal handler and only then passing the buck.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1766227