ureport: Generate GIR metadata
Misleadingly titled, this is a rather aggressive refactoring of the
code. In the end, it does achieve what the title suggests.

With this, we can rely on GObject Introspection and existing bindings
(PyGObject, for instance) instead of writing our own Python bindings by
hand, which is rather impossible to do without losing sanity when using
external libraries.

The refactoring bit simply turns the needed structs into boxed types,
hides the definitions behind public functions, so that we don’t poke at
the state and shoot ourselves in the foot^H^H^Hace. The tests were also
rewritten in plain C for debuggability.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>