ureport: Allow printf-like data attaching
Currently, it is only possible to attach string and integer data to
reports (without formatting a string beforehand), with the latter case
simply being a wrapper for the former.

This commit replaces the integer wrapper with a printf-like variadic
one should there ever be a need to attach more specific kinds of data.