Allow editing CHANGELOG.md before release commit
`$ make release-{major|minor|bug}` creates a new empty entry in CHANGELOG.md
but commits the change before the user has a chance to fill it, making
it necessary to ammend and re-tag the release commit. This commit opens
an editor to let the user edit the changelog in time.

Signed-off-by: Michal Fabik <mfabik@redhat.com>