reporter-gitlab: New reporter for GitLab
An implementation of a new reporting plugin for GitLab instances.
Basic functionality working so far: Allows to submit a new issue for
a component listed in the allowlist along with description of the
problem and basic info.

Two key features are still missing: duplicate detection and attachment
uploading. These will be implemented in future commits.