tests: Rewrite (de)compression test
Currently, the test relies on xz and lz4 binaries present to perform
the compression and is written using Autotest, neither of which are
ideal. Since 2fd7f5298a7602df623537797db39ba68336e7a4, any
compression-related matters are to be dealt with using the linked-in lz4
and lzma libraries and the latter issue is eliminated by rewriting the
test as a regular executable, to make debugging easier.