glib_support: Clean up properly after g_strsplit
Using the strings from the result of g_strsplit directly in a GList
causes a leak when the list is freed. This commit duplicates the strings
to be used as GList data (which is later freed along with the list
itself) and frees the original string vector with g_strfreev.

Also, fix leaks by avoiding the intermediate var in the loop and
g_strstrip'ing the vector members on the fly instead.

Signed-off-by: Michal Fabik <mfabik@redhat.com>