cli,lib,gui-wizard-gtk: Fix segfaults
Commit 7aba6e53b converted several pointers to GList into g_autolists
of GList. This is a mistake: the type parameter denotes the type of the
contained data, not of the container.

When the g_autolist'ed variable goes out of scope, the auto-cleanup
procedure attempts to free the data using the GList cleanup function,
which is g_list_free(). This leads to a segfault since the data are
usually strings allocated via malloc().