reporter-bugzilla: Fix bad error detection
reporter-bugzilla.c:925:25: warning: comparison of unsigned expression in ‘>= 0’ is always true [-Wtype-limits]
  925 |                 if (rtg >= 0 && rtg <= UINT_MAX && rating_str != endptr)
      |                         ^~

From g_ascii_strtoull documentation:
"
If the correct value would cause overflow, G_MAXUINT64 is returned,
and ERANGE is stored in errno. If the base is outside the valid
range, zero is returned, and EINVAL is stored in errno. If the
string conversion fails, zero is returned, and endptr returns
nptr (if endptr is non-NULL).
"