reporter-bugzilla: Reduce number of duplicate bug reports
Private bugs are problematic, because we might end up with many
duplicates. We simply don't know if the problem has been
reported already, if the previously reported bug is private.

We are going to open a public bug instead, but the bug description,
which might contain sensitive information, will be private.

We cannot make attachments private though. Attachments in Bugzilla
can be "Red Hat private", but we need users from certain groups
(like "fedora_contrib_private") to have access to them.
Therefore, we are going to create a separate supplementary bug
that will be private and will only contain the attachments.
This supplementary bug will be closed immediately, and the links
to the attachments will be added to the first public bug
in a private comment.

Signed-off-by: Michal Srb <michal@redhat.com>